+++ src/ControlPanel.DataAccess/Migrations/20250409053054_AddedLimitIdDirectories.cs
@@ -17,7 +17,7 @@ namespace ControlPanel.DataAccess.Migrations
                 {
                     Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
-                    LimitCode = table.Column<int>(type: "int", nullable: false),
+                    LimitCode = table.Column<long>(type: "bigint", nullable: false),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     DescriptionRu = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     DescriptionKg = table.Column<string>(type: "nvarchar(max)", nullable: false),
